We should make some UI changes to the Audit Screen
The title of the page should say Audit. And not Audit Summary. 
The label Search should be replaced with Filter. And we should use the
Filter icon that is used in the Reports section. 
The User drop down list should be alphabetically sorted
Event Occurrence should be replaced with Time

Using Data Filter